% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\name{annotate_loci}
\alias{annotate_loci}
\title{Generate transcript annotations}
\usage{
annotate_loci(
  ta,
  tx_archetypes,
  tx_fiveprime_dist,
  loci_gaps,
  loci_strands,
  telomere_buffer,
  seq_info
)
}
\arguments{
\item{ta}{a \link{transcript_archetypes-class} object}

\item{tx_archetypes}{a list of which transcript archetypes go to each locus grouped
by locus}

\item{tx_fiveprime_dist}{a list of transcript offsets from 5' end of the locus grouped
by locus}

\item{loci_gaps}{a vector of offsets between loci}

\item{loci_strands}{a vector of the strandedness of each loci}

\item{telomere_buffer}{regions at the head and tail of each chromosome that do not
contain transcripts}

\item{seq_info}{a \code{seqinfo} object used to determine chromosome lengths}
}
\value{
A \link[GenomicRanges]{GRanges-class}
}
\description{
Generate a \link[GenomicRanges]{GRanges-class} of transcript annotations
from a variety of inputs describing the loci, the archetypes, and their relative
offsets
}
