% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_processing.R
\name{filter_annotations}
\alias{filter_annotations}
\title{Filter annotations}
\usage{
filter_annotations(
  tx,
  min_distance = 20000,
  min_length = 3000,
  keep_chromosomes = NULL,
  keep_tx_biotype = NULL
)
}
\arguments{
\item{tx}{A \code{\link[GenomicRanges]{GRanges-class}} or
\code{\link[GenomicFeatures]{TxDb-class}}}

\item{min_distance}{minimum distance between candidate transcript and closest
transcript (strand agnostic).}

\item{min_length}{minimum length of candidate transcript}

\item{keep_chromosomes}{a character vector of chromosomes to select candidate
transcripts from. Defaults to all chromsomes in \code{tx} object}

\item{keep_tx_biotype}{a character vector of transcript types to select candidate
transcripts from. Defaults to all types.}
}
\value{
A list of vectors with each one corresponding to one set of bins and
each element of a vector corresponding to a bin
}
\description{
Gets a filtered candidate list of transcripts to use for simulation
}
