% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_processing.R
\name{inactive_regions}
\alias{inactive_regions}
\title{Get background regions}
\usage{
inactive_regions(
  tx,
  tx_buffer_length = 20000,
  min_length = 3000,
  keep_chromosomes = NULL
)
}
\arguments{
\item{tx}{A \code{\link[GenomicRanges]{GRanges-class}} or
\code{\link[GenomicFeatures]{TxDb-class}}}

\item{tx_buffer_length}{length of exclued region between inactive region and closest
transcript (default 40kb; strand agnostic).}

\item{min_length}{minimum length of candidate inactive region (default 3kb)}

\item{keep_chromosomes}{a character vector of chromosomes to select candidate
transcripts from. Defaults to all chromsomes in \code{tx} object}
}
\value{
A list of vectors with each one corresponding to one set of bins and
each element of a vector corresponding to a bin
}
\description{
Gets a filtered candidate list of background regions to use for
simulation
}
