% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{is.GRanges}
\alias{is.GRanges}
\title{is.GRanges}
\usage{
is.GRanges(x)
}
\arguments{
\item{x}{an object}
}
\value{
logical
}
\description{
Checks if object is a \code{\link[GenomicRanges]{GRanges-class}} object
}
