% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_checkers.R
\name{is.GRangesList}
\alias{is.GRangesList}
\title{is.GRangesList}
\usage{
is.GRangesList(x)
}
\value{
logical
}
\description{
Checks if object is a \code{\link[GenomicRanges]{GRangesList-class}} object
}
