% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_archives.R
\name{list_archetype_set}
\alias{list_archetype_set}
\title{List available archetype sets}
\usage{
list_archetype_set()
}
\value{
A character vector
}
\description{
Lists archetypes sets that are included in this package
}
