% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_distributions.R
\name{list_gtex_density}
\alias{list_gtex_density}
\title{List available density estimates from GTEx}
\usage{
list_gtex_density(pattern = NULL)
}
\arguments{
\item{pattern}{tissue name pattern to match}
}
\value{
character vector of tissue names
}
\description{
Lists available density estimates from GTEx for use with \code{rgtex}
}
