% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_archives.R
\name{load_archetype_set}
\alias{load_archetype_set}
\title{Load archetype set included in package}
\usage{
load_archetype_set(id = "celastrol_dukler2017")
}
\arguments{
\item{id}{the id of the archetype set to load}
}
\value{
A \code{transcript_archetypes} object
}
\description{
Load archetype set included in package. To see available sets run
\code{list_archetype_set()}.
}
