% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\name{loci_property_sim}
\alias{loci_property_sim}
\title{Simulate loci level properties}
\usage{
loci_property_sim(loci, n)
}
\arguments{
\item{loci}{A \link[GenomicRanges]{GRangesList-class} of transcripts grouped by loci}

\item{n}{the number of loci to simulate annotations for}
}
\value{
A list with two elements, the offsets between loci and the strands of the loci
}
\description{
Simulates the distance between loci and the strandedness of those loci
}
