% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\docType{class}
\name{profile_views-class}
\alias{profile_views-class}
\title{Class profile_views}
\description{
Class \code{profile_views} holds views of ranges on a dataset.
}
\section{Slots}{

\describe{
\item{\code{transcripts}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the query ranges}

\item{\code{profiles}}{a \code{\link[IRanges]{ViewsList-class}} object that holds the data
views}

\item{\code{idx}}{a lookup index used by acessors for the class}
}}

