% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{profile_views}
\alias{profile_views}
\title{extract regions from file}
\usage{
profile_views(bigwig_file, query_ranges)
}
\arguments{
\item{bigwig_file}{a string pointing to a bigwig file}

\item{query_ranges}{A \code{\link[GenomicRanges]{GRanges-class}} object}
}
\value{
A list of run length encodings
}
\description{
Extracts regions from a bigwig over a set of ranges from a
\code{\link[GenomicRanges]{GRanges-class}} object and build a
\code{\link{profile_views}} object
}
