% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_distributions.R
\name{rgtex}
\alias{rgtex}
\title{Sample from density estimates from GTEx}
\usage{
rgtex(n, tissue = "skeletal_muscle")
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the
number required.}

\item{tissue}{the tissue to simulate from (default: skeletal_muscle)}
}
\value{
Numeric vector of length \code{n}
}
\description{
Sample from probability density functions estimated from sex averaged
tissue specific transcript TPM from GTEx (with outliers removed).To view available
tissues use \code{list_gtex_density()}.
}
