% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_distributions.R
\name{rzexp}
\alias{rzexp}
\title{Simulate from a zero-inflated exponential distribution}
\usage{
rzexp(n, rate = 1, zero_weight = 0)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the
number required.}

\item{rate}{vector of rates}

\item{zero_weight}{probability of drawing a 0}
}
\value{
Numeric vector of length \code{n}
}
\description{
Zero-inflated exponential distribution
}
