% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\name{simulate_abundances}
\alias{simulate_abundances}
\title{Simulate abundance values}
\usage{
simulate_abundances(annotations, sampling_dist = rgtex, seed = NULL, ...)
}
\arguments{
\item{annotations}{a \link[GenomicRanges]{GRanges-class} of transcript annotations}

\item{sampling_dist}{a function that generates a vector of random numbers greater than
or equal to zero. It also must have as an argument \code{n} which specifies the number
of observations to draw. (A number of the random number generators from the stats
package fits all these requirements; eg. rlnorm, rpois, rgamma, etc.). Additionally,
you may design your own custom function to pass here as long as it meets these
criteria.}

\item{seed}{random seed for reproducible sampling}

\item{...}{any arguments to be passed to the \code{sampling_dist} function}
}
\value{
A \link[GenomicRanges]{GRanges-class} with a \code{score} column that holds
abundances
}
\description{
Take a \link[GenomicRanges]{GRanges-class} of transcript annotations
and return it annotated with abundance values of reach transcript
}
