% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\name{simulate_annotations}
\alias{simulate_annotations}
\title{Simulate annotations}
\usage{
simulate_annotations(
  ta,
  n,
  template,
  genome = NULL,
  tx_min_length = 3000,
  method = c("neighbor", "parametric"),
  min_loci_tx = 1,
  max_loci_tx = 40,
  seed = NULL
)
}
\arguments{
\item{ta}{a \link{transcript_archetypes-class} object}

\item{n}{the number of loci to simulate annotations for}

\item{template}{can be a \code{\link[GenomicRanges]{GRanges-class}} that provides
template annotations as a basis for simulating loci with realistic properties.}

\item{genome}{a genome name that will be used to retrieve chromosome lengths if
they are not already specified}

\item{tx_min_length}{do not simulate transcripts shorter than this length}

\item{method}{method to use for simulating intra-loci transcript properties. The two
methods are \code{neighbors} and \code{empirical}. More in information in details.}

\item{min_loci_tx}{minimum number of transcripts per loci}

\item{max_loci_tx}{maximum number of transcripts per loci}

\item{seed}{random seed to use for simulation}
}
\value{
A \link[GenomicRanges]{GRanges-class} of transcript annotations
}
\description{
Simulates per loci annotation
}
\details{
Stuff about methods
}
