% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate read data}
\usage{
simulate_data(ta, annotations, jitter = 10)
}
\arguments{
\item{ta}{a \link{transcript_archetypes-class} object}

\item{annotations}{a \link[GenomicRanges]{GRanges-class} of transcript annotations}

\item{jitter}{amount by which reads can be shifted to the left or right (default: 0)}
}
\value{
A \link[GenomicRanges]{GRanges-class}
}
\description{
Take a \link[GenomicRanges]{GRanges-class} of transcript annotations
and a \code{transcript_archetypes} object and simulate read counts.
}
