% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetype_archives.R
\name{store_reduced_archetype_bw}
\alias{store_reduced_archetype_bw}
\title{Save an archetype set}
\usage{
store_reduced_archetype_bw(
  bed_path,
  bigwig_plus,
  bigwig_minus,
  out_dir = ".",
  flank_length = 20000
)
}
\arguments{
\item{bed_path}{path to bed file that holds archetype regions}

\item{bigwig_plus}{path to bigwig for plus strand}

\item{bigwig_minus}{path to bigwig for minus strand}

\item{out_dir}{directory to save files in}

\item{flank_length}{length of region flanking those in bed file to save in the
subsetted bigwig files}
}
\description{
Save subset of data required to make an archetype set.
}
