% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{[,profile_views,numeric,missing,ANY-method}
\alias{[,profile_views,numeric,missing,ANY-method}
\alias{[}
\title{subsetting}
\usage{
\S4method{[}{profile_views,numeric,missing,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{\link{profile_views-class} object}

\item{i}{subset entries of views object}

\item{j}{unsupported}

\item{...}{just for compatibility with S4 method}

\item{drop}{unsupported}
}
\value{
A views object
}
\description{
Convenience subsetting for views object. Retrieves view based on index of
query range.
}
