% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{[,profile_views,numeric,missing-method}
\alias{[,profile_views,numeric,missing-method}
\title{subsetting}
\usage{
\S4method{[}{profile_views,numeric,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{i}{query range index}
}
\value{
A views object
}
\description{
Convenience subsetting for views object. Retrieves view based on index of
query range.
}
