% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\docType{class}
\name{transcript_archetypes-class}
\alias{transcript_archetypes-class}
\title{Class transcript_archetypes}
\description{
Class \code{transcript_archetypes} holds lists of empirical transcript archetypes
and the matched data. Empirical profiles for flanking regions (width is user defined)
are included in the archetypes. Synthetic genome annotations, count data, and
transcript abundances can be simulated from this object and exported to standard
genomic formats.
}
\section{Slots}{

\describe{
\item{\code{transcripts}}{a \code{\link[GenomicRanges]{GRanges-class}} that holds
all the transcript coordinates}

\item{\code{data}}{a list containing run-length encodings of counts for both the sense and
anti-sense strand for the transcripts and their flanking regions}

\item{\code{abundance_estimate}}{abundance estimates for each transcript}

\item{\code{configs}}{a list of the configuration options used to generate this object}
}}

