% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcript_archetypes.R
\name{view_archetype}
\alias{view_archetype}
\title{View archetype}
\usage{
view_archetype(ta, idx)
}
\arguments{
\item{ta}{a \link{transcript_archetypes-class} object}

\item{idx}{number id of archetype}
}
\description{
Plots archetype in sense/antisense orientation
}
